Large overhaul of phantom generation and cleanup
Added a truncation option to phantom generation to limit the
x, y, and z bounds of the phantom to be at most 2 cm out from the
nearest structures.  The buffer is also editeable to any number.

The mask generation and structure tallying has been fixed to
now actually ignore priority (tally voxels in 2 structures) and
mask media TARGET will be assigned for ALL contours which surround
a voxel.

Finally some general cleaup of the code.